Reorganize C++ code to avoid duplication in module bindings.
With the power of macros, we can move everything into the header files that can
be included in both its own compilation unit and in the module binding file.
This gives us the nice developer experience of modular compilation units, while
easily re-using the same definitions in the module bindings.